backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
statsmodels
joblib
python-dotenv
requests
google-generativeai
